+++ Formula1MyLive_Project/Formula1MyLive/Database/Model/ConstructorResult.cs
@@ -7,9 +7,9 @@ namespace Formula1MyLive.Database.Model
 {
 	public class ConstructorResult
 	{
-		public int Id { get; set; }
-		public int RaceId { get; set; }
-		public int ConstructorId { get; set; }
+		public Int16 Id { get; set; }
+		public Int16 RaceId { get; set; }
+		public Int16 ConstructorId { get; set; }
 		public int Points { get; set; }
 		public string Status { get; set; }
 	}
