+++ Formula1MyLive_Project/Formula1MyLive/Database/Model/ConstructorStanding.cs
@@ -7,12 +7,12 @@ namespace Formula1MyLive.Database.Model
 {
 	public class ConstructorStanding
 	{
-		public int Id { get; set; }
-		public int RaceId { get; set; }
-		public int ConstructorId { get; set; }
+		public Int16 Id { get; set; }
+		public Int16 RaceId { get; set; }
+		public Int16 ConstructorId { get; set; }
 		public int Points { get; set; }
-		public int Position { get; set; }
+		public Int16 Position { get; set; }
 		public string PositionText { get; set; }
-		public int Wins { get; set; }
+		public Int16 Wins { get; set; }
 	}
 }
