+++ Formula1MyLive_Project/Formula1MyLive/Database/Model/DriverStandingId.cs
@@ -8,11 +8,11 @@ namespace Formula1MyLive.Database.Model
 	public class DriverStanding
 	{
 		public int Id { get; set; }
-		public int RaceId { get; set; }
-		public int DriverId { get; set; }
+		public Int16 RaceId { get; set; }
+		public Int16 DriverId { get; set; }
 		public int Points { get; set; }
-		public int Position { get; set; }
+		public Int16 Position { get; set; }
 		public string PositionText { get; set; }
-		public int Wins { get; set; }
+		public Int16 Wins { get; set; }
 	}
 }
