+++ Formula1MyLive_Project/Formula1MyLive/Database/Model/Qualifying.cs
@@ -7,12 +7,12 @@ namespace Formula1MyLive.Database.Model
 {
 	public class Qualifying
 	{
-		public int Id { get; set; }
-		public int RaceId { get; set; }
-		public int DriverId { get; set; }
-		public int ConstructorId { get; set; }
-		public int Number { get; set; }
-		public int Position { get; set; }
+		public Int16 Id { get; set; }
+		public Int16 RaceId { get; set; }
+		public Int16 DriverId { get; set; }
+		public Int16 ConstructorId { get; set; }
+		public Int16 Number { get; set; }
+		public Int16 Position { get; set; }
 		public string Q1 { get; set; }
 		public string Q2 { get; set; }
 		public string Q3 { get; set; }
